Fix issue #33

+++ PortableLeagueApi.Interfaces/Static/Champion/ISpell.cs
@@ -6,7 +6,7 @@ namespace PortableLeagueApi.Interfaces.Static.Champion
     public interface ISpell : IApiModel
     {
         IList<IImage> AltImages { get; set; }
-        IList<int> Cooldown { get; set; }
+        IList<double> Cooldown { get; set; }
         string CooldownBurn { get; set; }
         IList<int> Cost { get; set; }
         string CostBurn { get; set; }
